requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database
pymongo==4.6.0

# LLM & AI Framework
crewai==0.28.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# NumPy (CRITICAL: Use stable version)
numpy==1.24.3

# NLP & Transformers
transformers==4.36.0
torch==2.1.2
tokenizers==0.15.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10
lxml==5.1.0

# Social Media APIs
tweepy==4.14.0
praw==7.8.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Additional dependencies
urllib3==2.1.0
certifi==2023.11.17
charset-normalizer==3.3.2